python/flask-postgres/requirements.txt
# Flask + PostgreSQL + OpenTelemetry Example
# Python 3.14+ | Dec 2025 Latest Versions

# Core Framework
flask>=3.1,<4.0
flask-sqlalchemy>=3.1,<4.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2

# Serialization
marshmallow>=4.1,<5.0
flask-marshmallow>=1.3,<2.0

# Utilities
python-slugify>=8.0

# Task Processing
celery>=5.6,<6.0
redis>=5.2,<6.0

# Authentication
PyJWT>=2.10,<3.0

# Configuration
python-dotenv>=1.0

# WSGI Server
gunicorn>=23.0

# OpenTelemetry (Dec 2025 latest)
opentelemetry-sdk>=1.39
opentelemetry-exporter-otlp>=1.39
opentelemetry-instrumentation-flask>=0.60b
opentelemetry-instrumentation-sqlalchemy>=0.60b
opentelemetry-instrumentation-celery>=0.60b
opentelemetry-instrumentation-redis>=0.60b
opentelemetry-instrumentation-logging>=0.60b

# Testing
pytest>=8.0
pytest-cov>=6.0
